1.修改时间精度为2位小数

+++ custom_widgets/lineeditwithvirtualnumerickeypad/icparameterconversion.cpp
@@ -38,7 +38,7 @@ QString ICParameterConversion::TransThisIntToThisText(int inputNum, int decimals
 
 int ICParameterConversion::TransTextToThisInt(const QString &numString, int decimals)
 {
-    int result = -1;
+    double result = -1;
         switch(decimals)
         {
         case 0:
